Create Data Integration
In this course, we explored the foundations of data integration techniques and their applications to big data. We utilized various tools, including Hadoop, Apache Spark, MapReduce, and cloud computing, to effectively manage and analyze data. This report aims to demonstrate the skills and knowledge acquired throughout the module by applying these techniques in practical scenarios.